+++ src/main/java/com/example/internship/dto/ProductSearchResult.java
@@ -13,5 +13,5 @@ public class ProductSearchResult {
     private List<ProductDto> products;
     private Integer pageNumber;
     private Integer pageSize;
-    private Integer totalProducts;
+    private Long totalProducts;
 }
